+++ src/node-lib/curriculum-api-2023/queries/unitListing/unitListing.query.test.ts
@@ -114,6 +114,7 @@ describe("unitListing()", () => {
       keyStageSlug: "ks1",
       keyStageTitle: "Key Stage 1",
       pathwayTitle: null,
+      pathwaySlug: null,
       examBoardSlug: null,
       examBoardTitle: null,
       subjectSlug: "maths",